+++ tests/xharness/Jenkins/TestTask/BuildProjectTask.cs
@@ -8,7 +8,7 @@ using System.Xml;
 using Xharness.Logging;
 using Xharness;
 
-namespace xharness.Jenkins.TestTask.TestTask
+namespace xharness.Jenkins.TestTask
 {
     abstract class BuildProjectTask : BuildToolTask
 	{
